requirements.txt
pyrogram>=2.0.0
tgcrypto
python-dotenv
sqlalchemy
alembic